RecepterDesign/requirements.txt
numpy
pandas
torch
networkx
scipy
pyro-ppl
pymc3
biopython
MDAnalysis
matplotlib
seaborn
snakemake
pyyaml
torch-geometric
e3nn
torch-scatter